allow to append a prefix and a suffix to the payload
the user can use --prefix "abcdfdafsjdlf" to generate a payload
prepended by "abcdfdafsjdlf". The same way, he is able to append after
the payload using --suffix.

Morever, he can specify --sfxnops 'n' to append 'n' NOP bytes directly
after the payload.

The user is informed via stderr of the offset of the NOP bytes and the
the offset of suffix.

Refs: #1